Running Mean, SD, Meadian
Write a program that reads a list of numbers and for each number outputs an estimate of the running mean, running standard deviation, and running median.  The input should be read from standard in, with one number per line.  For each line of input, the program should print to standard out the estimated running mean, running standard deviation, and running median.  That is, given the input

1

2

3

137.036

 

the program should output values close to

 

1,0,1

1.5,0.5,1.5

2,0.816,2

35.759,58.477,2.5

 

Some notes:

Using either the biased or unbiased estimate of the standard deviation is fine, but you should report a number even for the first step.
There are several ways to define the median for an even number of samples; you may use any of these definitions.
Output numbers at a reasonable precision.